Create CONTRIBUTING.md
As we get closer to go-live with documentation, we should align to a content structure that makes it clear where new content will live. This site will cover both Hubs and Spoke documentation, and we'll want to make it useful for audiences of end-users, room owners, content creators, and developers. 

This is a first attempt at a proposal, starting off a contributing file but I expect that this will grow to cover a greater scope. We may need to add and change categories, but wanted to get a starting point we could work from.