Add technical sections back in for technical pull requests
What: Adds the sections that are only relevant to technical pull requests back in (along with whether they're required or not), but with instructions not to include them for documentation pull requests.

Why: Not only documentation changes will be added to this repository, sometimes technical changes will be required, so we'll need people to provide the extra information for the technical pull requests while omitting it for the documentation pull requests where it isn't needed.