Update page descriptions
What: Updates page descriptions from admin-customizing-themes.md to hubs-cloud-content-cdn.md.

Why: To prevent accidental changes to page descriptions and provide better summaries for social media previews.

Notes:
This adds page description fields to the Docusaurus header so that Docusaurus will pull appropriate text from that for social media previews instead of using the first line of each document (which may or may not be suitable as a description, gives no indication that it is previewed in social media, and could be easily changed to something unsuitable for the social media previews by accident).