Update `admin-intro.md` description
What: Updates the description field for the `admin-intro.md` document to accurately reflect the uncommented content of the page and state where in the UI this is handled.

Why: The description was incorrect since it referenced parts of the document that were commented and not visible to people in the Docusaurus generated website.  Addresses review feedback.