Updates missing page descriptions
What: Adds page descriptions for 8 pages that were related to the setup of Hubs Cloud or Hubs Professional Plans.

Why: These descriptions were originally excluded because the pages are findable but not linked from within the current Hubs documentation. Further, Hubs Cloud and Hubs Professional Plans are defunct products.

Note: This commit is related to PR#223 and PR#224 which explain the importance of page descriptions.