Add ending periods to PR #224 document descriptions
What: Makes sure a period is at the end of every description for the files in PR #224.

Why: Social media link previews render the description exactly as it is provided, so to comply with standard English language rules a period should be added to the end of every description.  This was discussed and agreed to at the 2025-07-02 Hubs Documentation Session.

https://docs.google.com/document/d/1GhHG5KckJsb875BcMttKf2koBC6b-Ktq747NEE3kLiM/edit?tab=t.0#heading=h.ue7s28vohpl1

Note: Previously, the use of periods to end descriptions was inconsistent, so not all files needed a period added.