Modifies How much will this cost section
What: Modifies the How much will this cost section at the start of
the Beginners Guide. Makes clear total estimate cost per normal user
rises from $36US per month to $40US per month. Data backups are
not mentioned in this section as those were add-on services at DO, addressed
in the How to backup your Hubs instance doc.

Why: This section advices users of the costs upfront.
The upcoming upgrade to Hubs CE changes the data storage at DigitalOcean.
The cost per month for this storage is overall ~$2 less per month than the
DO charge for their Backups. We feel that users should be informed of
costs at the beginning of working on a Hubs Community Edition instance.

Note: the section, What you will get, is not addressed currently in this PR.
It might need to be added.

Review please and comments welcome.